requirements.txt
actionlib==1.13.2
angles==1.9.13
bondpy==1.8.6
camera-calibration==1.15.2
camera-calibration-parsers==1.12.0
catkin==0.8.8
controller-manager==0.19.2
controller-manager-msgs==0.19.2
cv-bridge==1.15.0
diagnostic-analysis==1.10.2
diagnostic-common-diagnostics==1.10.2
diagnostic-updater==1.10.2
dynamic-reconfigure==1.7.1
gazebo-plugins==2.9.1
gazebo-ros==2.9.1
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.14
image-geometry==1.15.0
interactive-markers==1.12.0
joint-state-publisher==1.15.0
joint-state-publisher-gui==1.15.0
laser-geometry==1.6.5
message-filters==1.15.8
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyserial==3.5
python-qt-binding==0.4.3
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
resource-retriever==1.12.6
rosbag==1.15.8
rosboost-cfg==1.15.7
rosclean==1.15.7
roscreate==1.15.7
rosgraph==1.15.8
roslaunch==1.15.8
roslib==1.15.7
roslint==0.12.0
roslz4==1.15.8
rosmake==1.15.7
rosmaster==1.15.8
rosmsg==1.15.8
rosnode==1.15.8
rosparam==1.15.8
rospy==1.15.8
rosservice==1.15.8
rostest==1.15.8
rostopic==1.15.8
rosunit==1.15.7
roswtf==1.15.8
rqt-action==0.4.9
rqt-bag==0.4.15
rqt-bag-plugins==0.4.15
rqt-console==0.4.11
rqt-dep==0.4.10
rqt-graph==0.4.14
rqt-gui==0.5.2
rqt-gui-py==0.5.2
rqt-image-view==0.4.16
rqt-launch==0.4.8
rqt-logger-level==0.4.11
rqt-moveit==0.5.9
rqt-msg==0.4.9
rqt-nav-view==0.5.7
rqt-plot==0.4.12
rqt-pose-view==0.5.10
rqt-publisher==0.4.9
rqt-py-common==0.5.2
rqt-py-console==0.4.9
rqt-reconfigure==0.5.3
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.11
rqt-runtime-monitor==0.5.8
rqt-rviz==0.6.1
rqt-service-caller==0.4.9
rqt-shell==0.4.10
rqt-srv==0.4.8
rqt-tf-tree==0.6.2
rqt-top==0.4.9
rqt-topic==0.4.12
rqt-web==0.4.9
rviz==1.14.1
sensor-msgs==1.13.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
topic-tools==1.15.8
xacro==1.14.4
